feat: Add DateFormat component to display formatted dates in MarkdownPostLayout
This commit adds the DateFormat component to the MarkdownPostLayout in order to display formatted dates for blog posts. The DateFormat component utilizes the formatDateIntl function to format the date according to the 'en-CA' locale. This enhancement improves the readability and consistency of the date display in the Astro-generated static site.