chore: Update npm dependencies to latest versions
This commit updates the npm dependencies in the `package.json` file to their latest stable versions. Specifically, it updates `@astrojs/mdx` to version `3.1.3`, `@astrojs/netlify` to version `5.4.0`, `@astrojs/react` to version `3.6.1`, and `astro` to version `4.13.0`. This ensures that the project is using the most up-to-date packages and takes advantage of any bug fixes or new features introduced in these versions.